+++ src/pages/task/finishTask/DetailTaskPage.tsx/FinishTask.tsx
@@ -191,7 +191,7 @@ function FinishTasks() {
           </Form.Item>
           <div className="row">
             <div className="col-lg">
-              <label>ภาพ Controller ควบคุม </label>
+              <label>ภาพหลักฐานการบิน</label>
               <br />
               <ImagCards
                 imageName={