+++ src/pages/task/finishTask/DetailTaskPage.tsx/ReviewTask.tsx
@@ -259,7 +259,7 @@ function ReviewTask() {
           </Form.Item>
           <div className="row">
             <div className="col-lg">
-              <label>ภาพ Controller ควบคุม </label>
+              <label>ภาพหลักฐานการบิน</label>
               <br />
               <ImagCards
                 imageName={